crOpenGL: Allow 64bit window ids for the parent window.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@22502 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/HostServices/SharedOpenGL/render/renderspu.h
@@ -251,7 +251,7 @@ typedef struct _VBOX_RENDERSPU_DESTROY_WINDOW {
 extern RenderSPU render_spu;
 
 /* @todo remove this hack */
-extern unsigned int render_spu_parent_window_id;
+extern uint64_t render_spu_parent_window_id;
 
 #ifdef CHROMIUM_THREADSAFE
 extern CRtsd _RenderTSD;
@@ -292,7 +292,7 @@ extern void RENDER_APIENTRY renderspuSwapBuffers( GLint window, GLint flags );
 #ifdef __cplusplus
 extern "C" {
 #endif
-DECLEXPORT(void) renderspuSetWindowId(unsigned int winId);
+DECLEXPORT(void) renderspuSetWindowId(uint64_t winId);
 DECLEXPORT(void) renderspuSetRootVisibleRegion(GLint cRects, GLint *pRects);
 #ifdef __cplusplus
 }
